requirements.txt
absl-py==0.11.0
alabaster==0.7.12
apturl==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.22
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gpg===1.13.1-unknown
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
ipykernel==5.4.2
ipyparallel==6.3.0
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
llvmlite==0.35.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.4
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mask-rcnn==2.1
matplotlib==3.3.3
mistune==0.8.4
monotonic==1.5
multitasking==0.0.9
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
networkx==2.5
nose==1.3.7
notebook==6.1.6
numba==0.52.0
numpy==1.19.5
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
openshot-qt==2.4.3
opt-einsum==3.3.0
packaging==20.3
pandas==1.2.0
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.4.11
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.15.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycocotools==2.0.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.7.3
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.17.3
pysmbc==1.0.15.6
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==5.0.1
QtPy==1.9.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.4.1
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.16.0
sip==4.19.21
six==1.15.0
snowballstemmer==2.1.0
speedtest-cli==2.1.2
Sphinx==3.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
system-service==0.3
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2021.2.26
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wxPython==4.1.1
wxwidgets==1.0.5
xkit==0.0.0
yarg==0.1.9
yfinance==0.1.64
zope.interface==4.7.1
